+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add `--show-ip-addresses` flag to `server list` command to optionally include IP addresses in command output.
 
+### Changed
+- Make `--family` parameter of `server firewall create` command optional to allow editing the default rules.
+
 ### Fixed
 - Complete shell input with uppercase letters (e.g., `Cap` to `CapitalizedName` will now work)
 - Display UUID of created template in `storage templatise` output.